Spanish mostly done except for white paper and needing to replace the flag.
I want the flag to be from the same icon set as the others. Web-devs, please stack up /images with as many as you can.

Also, is it possible to either have a 2 x 4 matrix of flags, or push the minilogo to the right a bit? The fifth flag covers it.